Add --purge-command for "purge"

+++ lib/expire/backups.rb
@@ -69,12 +69,16 @@ module Expire
       keep.length
     end
 
-    def purge(format)
+    def purge(format, &block)
       # expired.each { |backup| FileUtils.rm_rf(backup.id) }
       each do |backup|
         if backup.expired?
           format.before_purge(backup)
-          FileUtils.rm_rf(backup.path)
+          if block_given?
+            block.call(backup)
+          else
+            FileUtils.rm_rf(backup.path)
+          end
           format.after_purge(backup)
         else
           format.on_keep(backup)
